# back-end/docs/api/swagger.yaml
definitions:
  apperrors.AppError:
    properties:
      http_status:
        type: integer
      message:
        type: string
    type: object
...
  model.GenerateWalletsReq:
    properties:
      count:
        minimum: 1
        type: integer
      project_ids:
        items:
          type: integer
        type: array
    required:
    - count
    - project_ids
    type: object
  model.ImportWalletsReq:
    properties:
      private_keys:
        items:
          type: string
        type: array
      project_ids:
        items:
          type: integer
        type: array
    required:
    - private_keys
    - project_ids
    type: object
  model.MonitorWalletsReq:
    properties:
      before:
        description: base58 encoded transaction signature
        type: string
      page_size:
        type: integer
      wallet_ids:
        items:
          type: integer
        type: array
    type: object
  model.MonitorWalletsResp:
    properties:
      balance:
        example: 100.294
        type: number
      public_key:
        example: 88888888SOL88888888SOL88888888SOL8888888
        type: string
      transactions:
        items:
          $ref: '#/definitions/solanarpc.WalletTransaction'
        type: array
      wallet_id:
        example: 1234567890
        type: integer
    type: object
  model.PrivateKey:
    properties:
      private_key:
        example: 55bjLUoWhf2dFWnGEeLzCBdf5AuuJgWwRh7HghN2LoimnmjWz3qNgmni64x6nM3uTWqRNAET2cwef9pz21Zv4C2S
        type: string
    required:
    - private_key
    type: object
...
  model.Wallet:
    properties:
      created_at:
        example: "2025-10-01T03:12:20Z"
        type: string
      id:
        example: 1234567890
        type: integer
      private_key:
        example: 55bjLUoWhf2dFWnGEeLzCBdf5AuuJgWwRh7HghN2LoimnmjWz3qNgmni64x6nM3uTWqRNAET2cwef9pz21Zv4C2S
        type: string
      public_key:
        example: 88888888SOL88888888SOL88888888SOL8888888
        type: string
    type: object
...
  solanarpc.WalletTransaction:
    properties:
      amount:
        example: 100.294
        type: number
      timestamp:
        example: 1651366400
        type: integer
      transaction:
        example: "1234567890"
        type: string
    type: object
host: https://api-stage.tsunammi.io/
info:
  contact: {}
  description: This is the swagger specification for Tsunammi API.
  title: Tsunammi API
  version: 0.0.7
paths:
...
  /wallets/solana/{id}:
    get:
      description: Fetches the private key for a given wallet ID.
      operationId: fetch-private-key-by-id
      parameters:
      - description: Authentication token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Wallet ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.PrivateKey'
        "400":
          description: Bad Request (invalid public key)
          schema:
            $ref: '#/definitions/apperrors.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperrors.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apperrors.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrors.AppError'
      security:
      - BearerAuth: []
      summary: Get private key by wallet ID
      tags:
      - wallets
...
  /wallets/solana/generate:
    post:
      consumes:
      - application/json
      description: Generates a specified number of new Solana wallets and associates
        them with project IDs.
      operationId: generate-solana-wallets
      parameters:
      - description: Authentication token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Generation parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.GenerateWalletsReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: '#/definitions/model.Wallet'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperrors.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperrors.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrors.AppError'
      security:
      - BearerAuth: []
      summary: Generate Solana wallets
      tags:
      - wallets
  /wallets/solana/import:
    post:
      consumes:
      - application/json
      description: Imports existing Solana wallets using a list of private keys.
      operationId: import-solana-wallets
      parameters:
      - description: Authentication token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Import parameters with private keys
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.ImportWalletsReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: '#/definitions/model.Wallet'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperrors.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperrors.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrors.AppError'
      security:
      - BearerAuth: []
      summary: Import Solana wallets
      tags:
      - wallets
  /wallets/solana/import-file:
    post:
      description: Imports Solana wallets from a text file (.txt) containing private
        keys (one key per line).
      operationId: import-solana-wallets-from-file
      parameters:
      - description: Authentication token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Text file with private keys (one per line)
        in: formData
        name: file
        required: true
        type: file
      - description: JSON array of project IDs to associate wallets with.
        in: formData
        name: project_ids
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: '#/definitions/model.Wallet'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperrors.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperrors.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrors.AppError'
      security:
      - BearerAuth: []
      summary: Import Solana wallets from file
      tags:
      - wallets
  /wallets/solana/monitor:
    post:
      consumes:
      - application/json
      description: |-
        Starts monitoring of provided Solana wallets or updates monitoring settings.
        This endpoint allows adding wallets to the monitoring service to track their activity.
      operationId: monitor-solana-wallets
      parameters:
      - description: Authentication token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Monitoring parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.MonitorWalletsReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.MonitorWalletsResp'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperrors.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperrors.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrors.AppError'
      security:
      - BearerAuth: []
      summary: Monitor Solana wallets
      tags:
      - wallets
securityDefinitions:
  BearerAuth:
    description: Type 'Bearer' followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"